Set explicit table column sizing policy
When using the default table and column sizing policy, ImGui will assert if you pass in an init_width_or_weight to prevent you from mixing up width and weights.

Setting the column sizing policy explicitly to WidthStretch makes it more explicit that this is a weight, and doesn't change the behaviour.